requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.46
psycopg2-binary==2.9.9
pydantic-settings==2.13.1
anthropic==0.18.0
chromadb==0.4.22
sentence-transformers==2.5.1
trafilatura==1.8.0
httpx==0.27.0
pytest==8.0.0
python-dotenv==1.0.0